Move protocol ser/de tests into `core`
